Fixed lookup dogs

+++ hf_ueg/include/rhf.h
@@ -11,7 +11,7 @@ using namespace std;
 // RHF class
 class RHF {
 public:
-    RHF(const arma::mat &kinetic, const arma::vec &exchange, const size_t &n_elec, size_t &n_pw, const size_t &n_mom, arma::Mat<int> &plane_waves, arma::Mat<int> &momentum_transfer_vectors, const pair<arma::Mat<int>, arma::Mat<size_t>> &lookup_tables, const double &volume)
+    RHF(const arma::mat &kinetic, const arma::vec &exchange, const size_t &n_elec, size_t &n_pw, const size_t &n_mom, arma::Mat<int> &plane_waves, arma::Mat<int> &momentum_transfer_vectors, const pair<arma::Mat<int>, arma::Mat<int>> &lookup_tables, const double &volume)
         : kinetic(kinetic), interaction(exchange), n_elec(n_elec), n_pw(n_pw), n_mom(n_mom), plane_waves(plane_waves), momentum_transfer_vectors(momentum_transfer_vectors), lookup_tables(lookup_tables), volume(volume) {}
     // RHF-specific methods
     arma::mat guess_rhf(const string &guess_type);
@@ -27,7 +27,7 @@ private:
     size_t n_mom;
     arma::Mat<int> plane_waves;
     arma::Mat<int> momentum_transfer_vectors;
-    pair<arma::Mat<int>, arma::Mat<size_t>> lookup_tables;
+    pair<arma::Mat<int>, arma::Mat<int>> lookup_tables;
     double volume;
 };
